# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "cybersource",
    "contact": {},
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "variables": {}
    }
  ],
  "paths": {
...
    "/api/v1/merchants/register": {
      "post": {
        "tags": [
          "merchants"
        ],
        "summary": "register merchant",
        "operationId": "registermerchant",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/registermerchantrequest"
              },
              "example": {
                "bankType": "accessBank",
                "agid": "ACCESSNG0000002",
                "companyName": "merchant 8",
                "phoneNumber": "09065734222",
                "email": "merchant_80@email.com",
                "accountNumber": "0875314409",
                "bvn": "2635637431",
                "businessYears": 9,
                "websiteLink": "www.the-website.com",
                "address": "No 2 Luke Skywall Street",
                "primaryContactName": "The Primary Name",
                "primaryContactTelephone": "08984347322",
                "primaryContactPhoneNumber": "08033473333",
                "primaryContactEmail": "theprimary@email.com",
                "secondaryContactName": "The Secondary Name",
                "secondaryContactTelephone": "08044334444",
                "secondaryContactPhoneNumber": "08055544555",
                "secondaryContactEmail": "thesecondary@email.com"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    },
...
    "/api/v1/merchants/reset_password": {
      "post": {
        "tags": [
          "merchants"
        ],
        "summary": "reset merchant password",
        "operationId": "resetmerchantpassword",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/resetmerchantpasswordrequest"
              },
              "example": {
                "email": "valid_merchant_email"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "registermerchantrequest": {
        "title": "registermerchantrequest",
        "required": [
          "bankType",
          "agid",
          "companyName",
          "phoneNumber",
          "email",
          "accountNumber",
          "bvn",
          "businessYears",
          "websiteLink",
          "address",
          "primaryContactName",
          "primaryContactTelephone",
          "primaryContactPhoneNumber",
          "primaryContactEmail",
          "secondaryContactName",
          "secondaryContactTelephone",
          "secondaryContactPhoneNumber",
          "secondaryContactEmail"
        ],
        "type": "object",
        "properties": {
          "bankType": {
            "type": "string"
          },
          "agid": {
            "type": "string"
          },
          "companyName": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "accountNumber": {
            "type": "string"
          },
          "bvn": {
            "type": "string"
          },
          "businessYears": {
            "type": "integer",
            "format": "int32"
          },
          "websiteLink": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "primaryContactName": {
            "type": "string"
          },
          "primaryContactTelephone": {
            "type": "string"
          },
          "primaryContactPhoneNumber": {
            "type": "string"
          },
          "primaryContactEmail": {
            "type": "string"
          },
          "secondaryContactName": {
            "type": "string"
          },
          "secondaryContactTelephone": {
            "type": "string"
          },
          "secondaryContactPhoneNumber": {
            "type": "string"
          },
          "secondaryContactEmail": {
            "type": "string"
          }
        },
        "example": {
          "bankType": "accessBank",
          "agid": "ACCESSNG0000002",
          "companyName": "merchant 8",
          "phoneNumber": "09065734222",
          "email": "merchant_80@email.com",
          "accountNumber": "0875314409",
          "bvn": "2635637431",
          "businessYears": 9,
          "websiteLink": "www.the-website.com",
          "address": "No 2 Luke Skywall Street",
          "primaryContactName": "The Primary Name",
          "primaryContactTelephone": "08984347322",
          "primaryContactPhoneNumber": "08033473333",
          "primaryContactEmail": "theprimary@email.com",
          "secondaryContactName": "The Secondary Name",
          "secondaryContactTelephone": "08044334444",
          "secondaryContactPhoneNumber": "08055544555",
          "secondaryContactEmail": "thesecondary@email.com"
        }
      },
...
      "resetmerchantpasswordrequest": {
        "title": "resetmerchantpasswordrequest",
        "required": [
          "email"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          }
        },
        "example": {
          "email": "valid_merchant_email"
        }
      },
...
    }
  },
...
}